Made --part and --root deprecated and added --printer=none functionality.
cxxtest.py:
- reflects the changes in ../../../cxxtestgen.py
- CXXTEST_RUNNER got changed to CXXTEST_PYTHON
- CXXTEST_PRINTER got changed to CXXTEST_RUNNER because it governs the --runner
  option.

run_tests.sh:
- minor correction

printer_propagation/SConstruct
- environment variable change fix

cxxtestgen.py:
- deprecated --part, --root and --error-printer
- made --runner=none do the job of --part
- documentation updates to reflect the deprecation
- moved some of the code around so that now global emission checks are done as
  late as they can be. This simplifies writing new code because the global
  guards are already in place and it's one less thing to think about when
  writing.

+++ build_tools/SCons/test/run_tests.sh
@@ -34,7 +34,10 @@ function test() {
 }
 
 function cleanup() {
-    rm */.sconsign.dblite
+    if [[ "$ACTION" == "--clean" ]]
+    then
+        rm */.sconsign.dblite
+    fi
 }
 
 ACTION="--run"
